scripts/requirements.txt
kaggle>=1.6
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
python-dotenv>=1.0
httpx>=0.27
beautifulsoup4>=4.12
requests>=2.31
lxml>=4.9
